Ran Fastlane
$ fastlane produce
$ fastlane deliver init
$ fastlane deliver --force

iTunes Transporter output above ^
[13:15:05]: ERROR ITMS-90717: "Invalid App Store Icon. The App Store Icon in the asset catalog in 'ExpoKitApp.app' can't be transparent nor contain an alpha channel."
Return status of iTunes Transporter was 1: ERROR ITMS-90717: "Invalid App Store Icon. The App Store Icon in the asset catalog in 'ExpoKitApp.app' can't be transparent nor contain an alpha channel."
The call to the iTMSTransporter completed with a non-zero exit status: 1. This indicates a failure.